Separate sources and sinks
This separates some of the pieces a little bit more to how I expect they
will be used during batch generation.

* sources
  - no longer know about sinks
  - yield a seq of source locations, which have a name, and a list of
    tuples that represents the underlying data

* formatters
  - no longer know their extensions, but this was somewhat arbitrary.
    Seems like they didn't really care, and the edges can call things what
    they want anyway, so I just moved this into the sink.

* sinks
  - perform the io to write the formatted tile data out. I expect
    tilequeue will have an s3 implementation.

* generator
  - new concept that just threads the pieces together. I expect this to
    be the entry point for tilequeue.

Generators will assemble all the sources in the behavior that we'll
want, namely to iterate through all the sources and generate the
formatted data for a rawr tile grouping all at once. We could stream the
entire pipeline, but preventing writes if any querying or formatting
along the way failed seems preferable.