Limit intersections to polygons.
The intersection performed for the water and land polygon tables can produce geometry collections containing linestrings or points where the polygons touch (i.e: a zero area overlap) the bounding box.

This change makes sure that only polygonal results are returned, and a second filter ensures that the geometries are non-empty, as that might happen if a single linestring is filtered out.